# docs/api/openapi.yaml
openapi: 3.1.0
info:
  title: Mosaic API
  version: 1.0.0
  description: |
    Zero-knowledge encrypted photo gallery API.
    
    The server never sees plaintext photos or metadata—all encryption/decryption 
    happens client-side. The server stores encrypted blobs and facilitates 
    synchronization between clients.
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: /api
    description: API base path (behind reverse proxy)
security:
  - RemoteUser: []
...
paths:
...
  /albums/{albumId}/share-links:
    get:
      operationId: listShareLinks
      summary: List share links for an album
      description: Returns all share links created for the album. Owner only.
      tags: [Share Links]
      parameters:
        - $ref: '#/components/parameters/AlbumId'
        - $ref: '#/components/parameters/PageSkip'
        - $ref: '#/components/parameters/PageTake'
      responses:
        '200':
          description: List of share links
          headers:
            X-Pagination-Skip:
              $ref: '#/components/headers/PaginationSkip'
            X-Pagination-Take:
              $ref: '#/components/headers/PaginationTake'
            X-Pagination-Total-Count:
              $ref: '#/components/headers/PaginationTotalCount'
            X-Pagination-Has-More:
              $ref: '#/components/headers/PaginationHasMore'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PagedShareLinks'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'

    post:
      operationId: createShareLink
      summary: Create a new share link
      description: Creates a shareable link for anonymous album access. Owner only.
      tags: [Share Links]
      parameters:
        - $ref: '#/components/parameters/AlbumId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateShareLinkRequest'
      responses:
        '201':
          description: Share link created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ShareLinkResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
  /albums/{albumId}/share-links/with-secrets:
    get:
      operationId: listShareLinksWithSecrets
      summary: List share links with owner-encrypted secrets
      description: |
        Returns share links with their owner-encrypted secrets for epoch rotation.
        The secrets are encrypted with the owner's account key and can be used
        to derive wrapping keys for re-wrapping tier keys during rotation.
      tags: [Share Links]
      parameters:
        - $ref: '#/components/parameters/AlbumId'
        - $ref: '#/components/parameters/PageSkip'
        - $ref: '#/components/parameters/PageTake'
      responses:
        '200':
          description: List of share links with encrypted secrets
          headers:
            X-Pagination-Skip:
              $ref: '#/components/headers/PaginationSkip'
            X-Pagination-Take:
              $ref: '#/components/headers/PaginationTake'
            X-Pagination-Total-Count:
              $ref: '#/components/headers/PaginationTotalCount'
            X-Pagination-Has-More:
              $ref: '#/components/headers/PaginationHasMore'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PagedShareLinksWithSecrets'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
  /albums/{albumId}/share-links/{linkId}/expiration:
    patch:
      operationId: updateShareLinkExpiration
      summary: Update share-link expiration
      description: Updates expiration and max-use lifecycle settings for a share link; link secrets remain client-side.
      tags: [Share Links]
      parameters:
        - $ref: '#/components/parameters/AlbumId'
        - name: linkId
          in: path
          required: true
          schema: { type: string }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/UpdateLinkExpirationRequest' }
      responses:
        '200':
          description: Share-link expiration updated
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ShareLinkResponse' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
  /share-links/{linkId}:
    delete:
      operationId: revokeShareLink
      summary: Revoke a share link
      description: Permanently disables the share link. Owner only.
      tags: [Share Links]
      parameters:
        - name: linkId
          in: path
          required: true
          schema:
            type: string
            format: uuid
          description: Internal share link ID (not the public linkId)
      responses:
        '204':
          description: Share link revoked
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
  /share-links/{linkId}/keys:
    post:
      operationId: addShareLinkEpochKeys
      summary: Add epoch keys to a share link
      description: |
        Adds wrapped tier keys for new epochs during epoch rotation.
        Owner only. Cannot add keys to revoked links.
      tags: [Share Links]
      parameters:
        - name: linkId
          in: path
          required: true
          schema:
            type: string
            format: uuid
          description: Internal share link ID
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AddShareLinkEpochKeysRequest'
      responses:
        '200':
          description: Keys added successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  added:
                    type: integer
                    description: Number of new keys added
                  updated:
                    type: integer
                    description: Number of existing keys updated
        '400':
          $ref: '#/components/responses/BadRequest'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'

  # ===========================================================================
  # Anonymous Share Link Access (No Auth Required)
  # ===========================================================================
...
components:
...
  securitySchemes:
    RemoteUser:
      type: apiKey
      in: header
      name: Remote-User
      description: |
        User identity from trusted reverse proxy.
        Only accepted from configured trusted IP ranges.
    cookieAuth:
      type: apiKey
      in: cookie
      name: mosaic_session
      description: |
        LocalAuth session cookie. The backend sets this cookie as HttpOnly with Path=/api,
        SameSite=Lax for development/testing HTTP, and Secure + SameSite=Strict in production.

  # ===========================================================================
  # Parameters
  # ===========================================================================
  parameters:
    AlbumId:
      name: albumId
      in: path
      required: true
      description: Album UUID
      schema:
        type: string
        format: uuid
...
    PageSkip:
      name: skip
      in: query
      required: false
      description: Number of rows to skip before returning a page.
      schema:
        type: integer
        minimum: 0
        default: 0
    PageTake:
      name: take
      in: query
      required: false
      description: Page size. Values are clamped to the backend maximum of 100.
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 50
...
  headers:
    PaginationSkip:
      description: Normalized skip value applied by the backend.
      schema:
        type: integer
    PaginationTake:
      description: Normalized take value applied by the backend.
      schema:
        type: integer
    PaginationTotalCount:
      description: Total rows available for the query before paging.
      schema:
        type: integer
    PaginationHasMore:
      description: Whether another page exists after this response.
      schema:
        type: boolean

  # ===========================================================================
  # Responses
  # ===========================================================================
  responses:
    BadRequest:
      description: Invalid request
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/ProblemDetails'
...
    Forbidden:
      description: Insufficient permissions
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/ProblemDetails'
    NotFound:
      description: Resource not found
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/ProblemDetails'
...
  schemas:
...
    ProblemDetails:
      type: object
      description: RFC 7807 problem details error response.
      properties:
        type:
          type: string
        title:
          type: string
        status:
          type: integer
        detail:
          type: string
        instance:
          type: string
        correlationId:
          type: string
          description: Request correlation identifier when available.
      additionalProperties: true
    PagedResult:
      type: object
      required: [items, nextSkip]
      properties:
        items:
          type: array
          items: {}
        nextSkip:
          type: integer
          nullable: true
          description: Skip value for the next page, or null when no further page is available.
...
    PagedShareLinks:
      allOf:
        - $ref: '#/components/schemas/PagedResult'
        - type: object
          properties:
            items:
              type: array
              items:
                $ref: '#/components/schemas/ShareLinkResponse'
    PagedShareLinksWithSecrets:
      allOf:
        - $ref: '#/components/schemas/PagedResult'
        - type: object
          properties:
            items:
              type: array
              items:
                $ref: '#/components/schemas/ShareLinkWithSecretResponse'
...
    UpdateLinkExpirationRequest:
      type: object
      properties:
        expiresAt: { type: string, format: date-time, nullable: true }
        maxUses: { type: integer, minimum: 1, nullable: true }

    # -------------------------------------------------------------------------
    # Members
    # -------------------------------------------------------------------------
...
    ShareLinkResponse:
      type: object
      required:
        - id
        - linkId
        - accessTier
        - useCount
        - isRevoked
        - createdAt
      properties:
        id:
          type: string
          format: uuid
          description: Internal share link ID
        linkId:
          type: string
          description: Base64-encoded public link ID (derived from secret)
        accessTier:
          type: integer
          enum: [1, 2, 3]
          description: Access tier (1=thumb, 2=preview, 3=full)
        expiresAt:
          type: string
          format: date-time
          description: Optional expiration timestamp
        maxUses:
          type: integer
          description: Optional maximum uses limit
        useCount:
          type: integer
          description: Current number of accesses
        isRevoked:
          type: boolean
        createdAt:
          type: string
          format: date-time
    ShareLinkWithSecretResponse:
      type: object
      required:
        - id
        - linkId
        - accessTier
        - isRevoked
      properties:
        id:
          type: string
          format: uuid
        linkId:
          type: string
        accessTier:
          type: integer
          enum: [1, 2, 3]
        isRevoked:
          type: boolean
        ownerEncryptedSecret:
          type: string
          format: byte
          description: Base64-encoded owner-encrypted link secret
    CreateShareLinkRequest:
      type: object
      required:
        - accessTier
        - linkId
        - wrappedKeys
      properties:
        accessTier:
          type: integer
          enum: [1, 2, 3]
          description: Maximum access tier to grant
        linkId:
          type: string
          format: byte
          description: Base64-encoded link ID (16 bytes)
        ownerEncryptedSecret:
          type: string
          format: byte
          description: Base64-encoded owner-encrypted link secret
        expiresAt:
          type: string
          format: date-time
        maxUses:
          type: integer
          minimum: 1
        wrappedKeys:
          type: array
          items:
            $ref: '#/components/schemas/WrappedKeyRequest'
    WrappedKeyRequest:
      type: object
      required:
        - epochId
        - tier
        - nonce
        - encryptedKey
      properties:
        epochId:
          type: integer
        tier:
          type: integer
          enum: [1, 2, 3]
        nonce:
          type: string
          format: byte
          description: Base64-encoded 24-byte nonce
        encryptedKey:
          type: string
          format: byte
          description: Base64-encoded encrypted tier key
    AddShareLinkEpochKeysRequest:
      type: object
      required:
        - epochKeys
      properties:
        epochKeys:
          type: array
          items:
            $ref: '#/components/schemas/ShareLinkEpochKeyRequest'
    ShareLinkEpochKeyRequest:
      type: object
      required:
        - epochId
        - tier
        - nonce
        - encryptedKey
      properties:
        epochId:
          type: integer
        tier:
          type: integer
          enum: [1, 2, 3]
        nonce:
          type: string
          format: byte
          description: Base64-encoded 24-byte nonce
        encryptedKey:
          type: string
          format: byte
          description: Base64-encoded encrypted tier key

    # -------------------------------------------------------------------------
    # Anonymous Access Responses
    # -------------------------------------------------------------------------
...